+++ Changes.md
@@ -1,6 +1,11 @@
 1.7.x.x (relative to 1.7.0.0a9)
 =======
 
+Features
+--------
+
+- RandomPrimitiveVariable : Added new node for creating random per-face or per-vertex primitive variables with a variety of distributions.
+
 Fixes
 -----